+++ src/Symfony/Component/Console/Output/ConsoleSectionOutput.php
@@ -117,7 +117,7 @@ class ConsoleSectionOutput extends StreamOutput
         }
 
         if ($numberOfLinesToClear > 0) {
-            // Move cursor up n lines
+            // move cursor up n lines
             parent::doWrite(sprintf("\x1b[%dA", $numberOfLinesToClear), false);
             // erase to end of screen
             parent::doWrite("\x1b[0J", false);