+++ src/Symfony/Component/Console/Helper/Table.php
@@ -11,8 +11,10 @@
 
 namespace Symfony\Component\Console\Helper;
 
+use Symfony\Component\Console\Output\ConsoleSectionOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Exception\InvalidArgumentException;
+use Symfony\Component\Console\Exception\RuntimeException;
 
 /**
  * Provides helpers to display a table.
@@ -69,6 +71,8 @@ class Table
     private $columnWidths = array();
 
     private static $styles;
+
+    private $rendered = false;
 
     public function __construct(OutputInterface $output)
     {
@@ -252,6 +256,25 @@ class Table
         return $this;
     }
 
+    /**
+     * Adds a row to the table, and re-renders the table.
+     */
+    public function appendRow($row): self
+    {
+        if (!$this->output instanceof ConsoleSectionOutput) {
+            throw new RuntimeException(sprintf('Output should be an instance of "%s" when calling "%s".', ConsoleSectionOutput::class, __METHOD__));
+        }
+
+        if ($this->rendered) {
+            $this->output->clear($this->calculateRowCount());
+        }
+
+        $this->addRow($row);
+        $this->render();
+
+        return $this;
+    }
+
     public function setRow($column, array $row)
     {
         $this->rows[$column] = $row;
@@ -311,6 +334,7 @@ class Table
         $this->renderRowSeparator();
 
         $this->cleanup();
+        $this->rendered = true;
     }
 
     /**
@@ -443,6 +467,19 @@ class Table
                 }
             }
         });
+    }
+
+    private function calculateRowCount(): int
+    {
+        $numberOfRows = count(iterator_to_array($this->buildTableRows(array_merge($this->headers, array(new TableSeparator()), $this->rows))));
+
+        if ($this->headers) {
+            ++$numberOfRows; // Add row for header separator
+        }
+
+        ++$numberOfRows; // Add row for footer separator
+
+        return $numberOfRows;
     }
 
     /**