+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 -----
 
  * added option to run suggested command if command is not found and only 1 alternative is available
- * added option to modify console output and print multiple modifyable sections
+ * added option to modify console output and print multiple modifiable sections
 
 4.0.0
 -----